+++ web/packages/media/file-metadata.ts
@@ -3,10 +3,10 @@ import { authenticatedRequestHeaders, ensureOk } from "@/base/http";
 import { apiURL } from "@/base/origins";
 import { type Location } from "@/base/types";
 import { type EnteFile, type FilePublicMagicMetadata } from "@/media/file";
-import { mergeMetadata1 } from "@/new/photos/utils/file";
 import { ensure } from "@/utils/ensure";
 import { nullToUndefined } from "@/utils/transform";
 import { z } from "zod";
+import { mergeMetadata1 } from "./file";
 import { FileType } from "./file-type";
 
 /**
